Project refactored and Http codes implemented

+++ src/main/java/com/spring/rest/controller/StudentController.java
@@ -3,6 +3,8 @@ package com.spring.rest.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,8 +13,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.spring.jdbc.model.Student;
 import com.spring.rest.jdbc.dao.StudentDAO;
+import com.spring.rest.model.Student;
 
 @RestController
 public class StudentController {
@@ -21,33 +23,71 @@ public class StudentController {
 		StudentDAO studentDAO;
 		
 		@GetMapping("/students")
-		List<Student> getAllStudents() {
-			return studentDAO.getAllStudents();
+		ResponseEntity<List<Student>> getAllStudents() {
+			try {
+				List<Student> result = studentDAO.getAllStudents();
+				if(!result.isEmpty()) {
+					return ResponseEntity.status(HttpStatus.OK).body(result);
+				}
+				return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+				
+			} catch (IllegalStateException e) {
+				return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+			}
 		}
 		
-		@PostMapping("/addStudent")
-		Student addStudent(@RequestBody Student newStudent) {
-			return studentDAO.addStudent(newStudent);
+		@PostMapping("/students/add")
+		ResponseEntity<Student> addStudent(@RequestBody Student newStudent) {
+			try {
+				studentDAO.addStudent(newStudent);
+				return ResponseEntity.status(HttpStatus.CREATED).build();
+			} catch (Exception e) {
+				e.printStackTrace();
+				return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+			}
 		}
 		
 		@GetMapping("/students/{id}")
-		Student findStudent(@PathVariable Integer id) throws Exception {
+		ResponseEntity<Student> findStudent(@PathVariable Integer id) throws Exception {
 			Student student = studentDAO.findStudentById(id);
-			
-			if(student == null) {
-				throw new Exception("No Student found for id: " + id);
+			try {
+				if(student != null) {
+					return ResponseEntity.status(HttpStatus.OK).body(student);
+				}
+				return ResponseEntity.status(HttpStatus.NOT_FOUND).build();				
+			} catch (Exception e) {
+				e.getMessage();
+				return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
 			}
-			
-			return student;
 		}
 		
 		@PutMapping("/students/{id}")
-		Student updateStudent(@RequestBody Student student, @PathVariable Integer id) {
-			return studentDAO.updateStudent(id, student);
+		ResponseEntity<Student> updateStudent(@RequestBody Student student, @PathVariable Integer id) {
+			try {
+				if(studentDAO.findStudentById(id) != null) {
+					Student result = studentDAO.updateStudent(id, student);
+					return ResponseEntity.status(HttpStatus.OK).body(result);
+				}
+				return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+				
+			} catch (Exception e) {
+				e.getMessage();
+				return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+			}
+
 		}
 		
 		@DeleteMapping("/students/{id}")
-		void deleteStudent(@PathVariable Integer id) {
-			studentDAO.deleteStudent(id);
+		ResponseEntity<Void> deleteStudent(@PathVariable Integer id) {
+			try {
+				if(studentDAO.findStudentById(id) != null) {
+					studentDAO.deleteStudent(id);
+					return ResponseEntity.status(HttpStatus.OK).build();
+				}
+				return ResponseEntity.status(HttpStatus.NOT_FOUND).build();				
+			}catch (Exception e) {
+				e.getMessage();
+				return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+			}
 		}
 }
